requirements.txt
opencv-python>=4.8.1.78
numpy>=1.26.1
pandas
scikit-learn>=1.3.2
tensorflow>=2.14.0
mtcnn>=0.1.1

keras-facenet>=0.3.2

sqlalchemy-pgvector
pgvector

Flask>=3.0.0
fastapi>=0.106.0 
uvicorn>=0.25.0 